+++ src/utils/remove-0x.ts
@@ -1,3 +1,0 @@
-export function remove0x(hex: string): string {
-  return hex.replace(/^0x/, '');
-}