+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `BlockResolverWroker`
+- db migrations
+- retries on timeout
+- do first call for leaves to block minter
+
+### Changed
+- pull consensus data from blockchain
+
+### Fixed
+- do not stuck on blocks when validators calls for data fail
 
 ## [1.0.1] - 2021-05-25
 ### Changed